+++ .env.tmpl
@@ -5,12 +5,11 @@ export AWS_SECRET_KEY="Your AWS SECRET KEY"
 export PROJECT_NAME="Prefix all of the resource"
 export COMMIT_HASH="Your Git Commit's Hash"
 
-export DOMAIN_NAME="Your Domain's name"
-export DOMAIN_SERVER_NAME_1="Your Domain's Server Name1"
-export DOMAIN_SERVER_NAME_2="Your Domain's Server Name2"
-export DOMAIN_SERVER_NAME_3="Your Domain's Server Name3"
-export DOMAIN_SERVER_NAME_4="Your Domain's Server Name4"
 export HOSTED_ZONE_ID="Your Domain's Hosted Zone ID"
+export SUB_DOMAIN_NAME="SubDomain Name That You Want it
+       ex) Your Input : "test"
+           Domain Name That You Already Have : mydomin.com
+           Final subdomain name to be registered : test.mydomain.com
 
 export EB_SOLUTION_STACK="EC2's Application Runtime"
 export SOURCE_BUNDLE_PATH="The local location of the JAR file running on EC2. (The reason why this is necessary is because when tf deploy is performed, the JAR file to be deployed to S3 is uploaded. So S3 needs to know where the JAR file is located.)"
